% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameTree.R
\name{gameTree}
\alias{gameTree}
\title{Plot a tree diagram for a simple 2x2 game theory model.}
\usage{
gameTree(
  players = c("Player 1", "Player 2"),
  strategies = c("Collude", "Defect"),
  payoffs1 = NULL,
  payoffs2 = NULL
)
}
\arguments{
\item{players}{is a character vector of name labels for the players in each stage (default is \code{c("Player 1", "Player 2")}.}

\item{strategies}{is a character vector of name labels for the strategies (default is \code{c("Collude", "Defect")}.}

\item{payoffs1}{is a numerical vector of length 4 listing the payoffs of player 1 (default is \code{c(0, 0, 0, 0)})}

\item{payoffs2}{is a numerical vector of length 4 listing the payoffs of player 2 (default is \code{c(0, 0, 0, 0)})}
}
\value{
A ggplot object with a 2x2 game tree.
}
\description{
Plot a tree diagram for a simple 2x2 game theory model.
}
\details{
\code{gameTree} draws a game tree with specified players ("Player 1" and "Player 2) who sequentially choose from the same strategies ("Collude" or "Defect") to earn payoffs.
}
