% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solnp2.R
\name{solnp2}
\alias{solnp2}
\title{Maximize a function with `solnp()`..}
\usage{
solnp2(
  pars,
  fun,
  eqfun = NULL,
  eqB = NULL,
  ineqfun = NULL,
  ineqLB = NULL,
  ineqUB = NULL,
  max = FALSE,
  control = list(),
  ...
)
}
\arguments{
\item{pars}{The starting parameter vector.}

\item{fun}{The main function which takes as first argument the parameter vector and returns a single value.}

\item{eqfun}{(Optional) The equality constraint function returning the vector of evaluated equality constraints.}

\item{eqB}{(Optional) The equality constraints.}

\item{ineqfun}{(Optional) The inequality constraint function returning the vector of evaluated inequality constraints.}

\item{ineqLB}{(Optional) The lower bound of the inequality constraints.}

\item{ineqUB}{(Optional) The upper bound of the inequality constraints.}

\item{control}{(Optional) The control list of optimization parameters. See below for details.}

\item{LB}{(Optional) The lower bound on the parameters.}

\item{UB}{(Optional) The upper bound on the parameters.}

\item{type}{is a logical value indicating whether to clear the main tutorial or practice.}
}
\description{
This is a wrapper function adds an option to `Rsolnp::solnp()` to switch to maximization instead of minimization..
}
